[IMP] project_stock: Add analytic date
The possibility of set a specific date in project or task has been added
so that if it is set it will be used in analytic items linked to materials.

TT38988

[UPD] Update project_stock.pot

project_stock 14.0.1.1.0

Update translation files

Updated by "Update PO files to match POT (msgmerge)" hook in Weblate.

Translation: project-14.0/project-14.0-project_stock
Translate-URL: https://translation.odoo-community.org/projects/project-14-0/project-14-0-project_stock/